Updated readme and refactored tests slightly
