--- AgentDiscoveries-Frontend/app/end_to_end_tests/helpers/setupTests.js
@@ -1,21 +1,0 @@
-import {configure} from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-var jsdom = require('jsdom');
-
-const {JSDOM} = jsdom;
-configure({adapter: new Adapter()});
-
-var exposedProperties = ['window', 'navigator', 'document'];
-const {document} = (new JSDOM('')).window;
-global.document = document;
-Object.keys(document.defaultView).forEach((property) => {
-    if (typeof global[property] === 'undefined') {
-        exposedProperties.push(property);
-        global[property] = document.defaultView[property];
-    }
-});
-
-global.navigator = {
-    userAgent: 'node.js'
-};
